# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Shelter Center API
  version: 1.0.0
  description: "유기동물 보호소 운영 API (Animals/Adoptions/Donations/Medical/Volunteers/Stats)"
servers:
  - url: http://localhost:8080
    description: local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [timestamp, path, status, code, message]
      properties:
        timestamp: { type: string, format: date-time }
        path: { type: string }
        status: { type: integer }
        code: { type: string }
        message: { type: string }
        details: { type: object, nullable: true }
    PageResponse:
      type: object
      required: [content, page, size, totalElements, totalPages, sort]
      properties:
        content:
          type: array
          items: {}
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer }
        totalPages: { type: integer }
        sort: { type: string }
...
    Adoption:
      type: object
      properties:
        id: { type: integer }
        animalId: { type: integer }
        applicantName: { type: string }
        applicantPhone: { type: string, nullable: true }
        status:
          type: string
          enum: [REQUESTED, APPROVED, CONFIRMED, CANCELED]
        requestedAt: { type: string, format: date-time }
        approvedAt: { type: string, format: date-time, nullable: true }
        adoptedAt: { type: string, format: date-time, nullable: true }
        canceledAt: { type: string, format: date-time, nullable: true }
        cancelReason: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
paths:
...
  /api/adoptions/adoptions:
    post:
      tags: [Adoptions]
      summary: Create adoption request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [animalId, applicantName]
              properties:
                animalId: { type: integer }
                applicantName: { type: string, maxLength: 50 }
                applicantPhone: { type: string, nullable: true, maxLength: 20 }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Adoption" }
        "409":
          description: state conflict
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422":
          description: validation failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

    get:
      tags: [Adoptions]
      summary: List adoptions (filter/search/sort/page)
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [REQUESTED, APPROVED, CONFIRMED, CANCELED]
        - in: query
          name: animalId
          schema: { type: integer }
        - in: query
          name: keyword
          schema: { type: string, description: "applicantName contains" }
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
        - in: query
          name: page
          schema: { type: integer, default: 0, minimum: 0 }
        - in: query
          name: size
            # validators 기준 1~50 가정
          schema: { type: integer, default: 20, minimum: 1, maximum: 50 }
        - in: query
          name: sort
          schema: { type: string, default: "requestedAt,DESC" }
      responses:
        "200":
          description: paged adoptions
          content:
            application/json:
              schema:
                allOf: [{ $ref: "#/components/schemas/PageResponse" }]
                properties:
                  content:
                    type: array
                    items: { $ref: "#/components/schemas/Adoption" }
        "400":
          description: invalid query
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/adoptions/adoptions/{id}:
    get:
      tags: [Adoptions]
      summary: Get adoption by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: adoption
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Adoption" }
        "404":
          description: not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "400":
          description: invalid id
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/adoptions/adoptions/pending:
    get:
      tags: [Adoptions]
      summary: List pending adoption requests (REQUESTED)
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 0, minimum: 0 }
        - in: query
          name: size
          schema: { type: integer, default: 20, minimum: 1, maximum: 50 }
      responses:
        "200":
          description: paged requested adoptions
          content:
            application/json:
              schema:
                allOf: [{ $ref: "#/components/schemas/PageResponse" }]
                properties:
                  content:
                    type: array
                    items: { $ref: "#/components/schemas/Adoption" }
  /api/adoptions/animals/{id}/adoptions:
    get:
      tags: [Adoptions]
      summary: List adoptions by animal id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, description: "animalId" }
        - in: query
          name: page
          schema: { type: integer, default: 0, minimum: 0 }
        - in: query
          name: size
          schema: { type: integer, default: 20, minimum: 1, maximum: 50 }
      responses:
        "200":
          description: paged adoptions for animal
          content:
            application/json:
              schema:
                allOf: [{ $ref: "#/components/schemas/PageResponse" }]
                properties:
                  content:
                    type: array
                    items: { $ref: "#/components/schemas/Adoption" }
        "404":
          description: animal not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/adoptions/adoptions/{id}/approve:
    patch:
      tags: [Adoptions]
      summary: Approve adoption (ADMIN)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: approved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Adoption" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Forbidden (role)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: state conflict
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/adoptions/adoptions/{id}/confirm:
    patch:
      tags: [Adoptions]
      summary: Confirm adoption (adoptedAt required)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [adoptedAt]
              properties:
                adoptedAt: { type: string, format: date-time }
      responses:
        "200":
          description: confirmed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Adoption" }
        "422":
          description: validation failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: state conflict
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/adoptions/adoptions/{id}/cancel:
    patch:
      tags: [Adoptions]
      summary: Cancel adoption (cancelReason required)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cancelReason]
              properties:
                cancelReason: { type: string, maxLength: 255 }
      responses:
        "200":
          description: canceled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Adoption" }
        "422":
          description: validation failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: state conflict
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  ########################################################
  # Donations (mounted: /api/donations)  ※ 내부가 /donations...
  ########################################################
...